Teste API Projeto/requirements.txt
asttokens==2.2.1
backcall==0.2.0
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
decorator==5.1.1
distlib==0.3.6
executing==1.2.0
filelock==3.9.0
Flask==2.2.3
Flask-Cors==3.0.10
idna==3.4
ipython==8.10.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.0.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pure-eval==0.2.2
pwinput==1.0.3
Pygments==2.14.0
python-dotenv==0.21.1
requests==2.28.2
six==1.16.0
stack-data==0.6.2
traitlets==5.9.0
urllib3==1.26.14
virtualenv==20.19.0
wcwidth==0.2.6
Werkzeug==2.2.3
